+++ lib/eeny-meeny/models/variation.rb
@@ -8,13 +8,5 @@ module EenyMeeny
       @weight = weight
       @options = options
     end
-
-    def marshal_dump
-      [@id, { name: @name, weight: @weight, **@options }]
-    end
-
-    def marshal_load(array)
-      send :initialize, array[0], **array[1]
-    end
   end
 end