+++ lib/eeny-meeny/routing/experiment_constraint.rb
@@ -10,10 +10,10 @@ module EenyMeeny
     end
 
     def matches?(request)
-      return false unless @experiment.active?
-      cookie = EenyMeeny::Cookie.read(request.cookie_jar[EenyMeeny::Cookie.cookie_name(@experiment)])
-      return false if cookie.nil? # Not participating in experiment
-      (@variation_id.nil? || @variation_id == cookie[:variation].id)
+      return false unless !@experiment.nil? && @experiment.active?
+      participant_variation_id = EenyMeeny::Cookie.read(request.cookie_jar[EenyMeeny::Cookie.cookie_name(@experiment)])
+      return false if participant_variation_id.nil? # Not participating in experiment
+      (@variation_id.nil? || @variation_id == participant_variation_id)
     end
   end
 end